+++ net-spider/ChangeLog.md
@@ -6,6 +6,7 @@
   in `GraphML.Writer` module are moved to `GraphML.Attribute` module.
 * Add `ToAttributes` instance to `TimeZone`.
 * Add `FromJSON` and `ToJSON` instances to `AttributeValue`.
+* Add `attributesToAeson` function.
 
 ### Timestamp module